+++ src/Libraries/SmartStore.Services/Media/IPictureService.cs
@@ -26,6 +26,16 @@ namespace SmartStore.Services.Media
 		/// <param name="equalPictureId">Id of equal picture if any</param>
 		/// <returns>The picture binary for <c>path</c> when no picture equals in the sequence, <c>null</c> otherwise.</returns>
 		byte[] FindEqualPicture(string path, IEnumerable<Picture> productPictures, out int equalPictureId);
+
+		/// <summary>
+		/// Finds an equal picture by comparing the binary buffer
+		/// </summary>
+		/// <param name="pictureBinary">Binary picture data</param>
+		/// <param name="productPictures">The sequence of product pictures to seek within for duplicates</param>
+		/// <param name="equalPictureId">Id of equal picture if any</param>
+		/// <returns>The picture binary for <c>path</c> when no picture equals in the sequence, <c>null</c> otherwise.</returns>
+		byte[] FindEqualPicture(byte[] pictureBinary, IEnumerable<Picture> productPictures, out int equalPictureId);
+
         
         /// <summary>
         /// Gets the loaded picture binary depending on picture storage settings